changed cross to another page
